🎨 Folder structure improved and forms option added
