+++ config/envs/unit_testing.py
@@ -1,3 +1,4 @@
+# flake8 : noqa
 """Flask Unit Testing Environment Configuration."""
 from os import environ
