+++ db/exc/__init__.py
@@ -1,0 +1,3 @@
+from .application import ApplicationError
+
+__all__ = [ApplicationError]
