+++ db/exc/application.py
@@ -1,0 +1,10 @@
+class ApplicationError(Exception):
+    """Exception raised for errors in Application management
+
+    Attributes:
+        message -- explanation of the error
+    """
+
+    def __init__(self, message="Sorry, there was a problem, please try later"):
+        self.message = message
+        super().__init__(self.message)
