+++ db/models/__init__.py
@@ -1,2 +1,4 @@
-from .applications import Applications
+from .application import Applications
 from .forms import Forms
+
+__all__ = [Applications, Forms]
