+++ db/models/forms/__init__.py
@@ -1,0 +1,1 @@
+from .forms import Forms
