--- db/models/language.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class Language(Enum):
-    en = 0
-    cy = 1
