+++ db/queries/__init__.py
@@ -1,1 +1,31 @@
+from .application import create_application
 from .application import get_application
+from .application import get_count_by_status
+from .application import search_applications
+from .application import submit_application
+from .form import add_new_forms
+from .form import get_form
+from .form import get_forms_by_app_id
+from .reporting import export_json_to_csv
+from .reporting import get_general_status_applications_report
+from .reporting import get_report_for_all_applications
+from .reporting import get_report_for_application
+from .updating import update_application_and_related_form
+from .updating import update_form
+
+__all__ = [
+    create_application,
+    get_application,
+    get_count_by_status,
+    search_applications,
+    submit_application,
+    add_new_forms,
+    get_form,
+    get_forms_by_app_id,
+    export_json_to_csv,
+    get_general_status_applications_report,
+    get_report_for_all_applications,
+    get_report_for_application,
+    update_application_and_related_form,
+    update_form,
+]
