+++ external_services/data.py
@@ -1,74 +1,15 @@
 import json
 import os
-from operator import itemgetter
 from typing import List
 from urllib.parse import urlencode
 
 import requests
 from config import Config
-from external_services.models.account import Account
-from external_services.models.fund import Fund
-from external_services.models.round import Round
+from .models.account import Account
+from .models.fund import Fund
+from .models.round import Round
 from flask import abort
 from flask import current_app
-
-
-def get_forms_from_form_config(form_config, fund_id, round_id, language):
-    mint_form_list = set()
-    forms_config = form_config.get(":".join([fund_id, round_id]))
-    for form_config in forms_config:
-        for form in form_config['ordered_form_names_within_section']:
-            mint_form_list.add(form[language])
-    return mint_form_list
-
-class ApplicationHelpers:
-    @staticmethod
-    def get_blank_forms(fund_id: str, round_id: str, language: str):
-        """
-        Get the list of forms required to populate a blank
-        application for a fund round
-
-        Args:
-            fund_id: (str) The id of the fund
-            round_id: (str) The id of the fund round
-
-        Returns:
-            A list of json forms to populate the form
-        """
-        fund = get_fund(fund_id)
-        fund_round = get_round(fund_id, round_id)
-        if fund and fund_round:
-            form_config = Config.FORMS_CONFIG_FOR_FUND_ROUND
-            forms = get_forms_from_form_config(form_config, fund_id, round_id, language)
-            if not forms:
-                raise Exception(
-                    f"Could not find forms for {fund_id} - {round_id}"
-                )
-            return forms.copy()
-        raise Exception(
-            f"Could not find fund round for {fund_id} - {round_id}  in fund"
-            " store."
-        )
-
-    def order_applications(applications, order_by, order_rev):
-        """
-        Returns a list of ordered applications
-        """
-        if order_by and order_by in [
-            "id",
-            "status",
-            "account_id",
-            "assessment_deadline",
-            "started_at",
-            "last_edited",
-        ]:
-            applications = sorted(
-                applications,
-                key=itemgetter(order_by),
-                reverse=int(order_rev),
-            )
-
-        return applications
 
 
 def get_data(endpoint: str, params: dict = None):
@@ -90,9 +31,7 @@ def get_data(endpoint: str, params: dict = None):
         current_app.logger.info(f"Fetching data from '{endpoint}'.")
         data = get_remote_data(endpoint, params)
     if data is None:
-        current_app.logger.error(
-            f"Data request failed, unable to recover: {endpoint}"
-        )
+        current_app.logger.error(f"Data request failed, unable to recover: {endpoint}")
         return abort(500)
     return data
 
@@ -148,12 +87,9 @@ def get_funds() -> List[Fund] | None:
 
 
 def get_fund(fund_id: str) -> Fund | None:
-    endpoint = Config.FUND_STORE_API_HOST + Config.FUND_ENDPOINT.format(
-        fund_id=fund_id
-    )
+    endpoint = Config.FUND_STORE_API_HOST + Config.FUND_ENDPOINT.format(fund_id=fund_id)
     current_app.logger.info(f"Request made to {endpoint}")
     response = get_data(endpoint)
-    print(response)
     if response is None:
         current_app.logger.info("Request to fund store returned None")
     fund = Fund.from_json(response)
@@ -176,9 +112,8 @@ def get_round(fund_id: str, round_id: str) -> Round | None:
     """
     Gets round from round store api using round_id if given.
     """
-    round_endpoint = (
-        Config.FUND_STORE_API_HOST
-        + Config.FUND_ROUND_ENDPOINT.format(fund_id=fund_id, round_id=round_id)
+    round_endpoint = Config.FUND_STORE_API_HOST + Config.FUND_ROUND_ENDPOINT.format(
+        fund_id=fund_id, round_id=round_id
     )
     round_response = get_data(round_endpoint)
     if round_response and "id" in round_response:
@@ -211,3 +146,12 @@ def get_account(email: str = None, account_id: str = None) -> Account | None:
 
     if response and "account_id" in response:
         return Account.from_json(response)
+
+
+def get_round_name(fund_id, round_id):
+    response = get_data(
+        Config.FUND_STORE_API_HOST
+        + Config.FUND_ROUND_ENDPOINT.format(fund_id=fund_id, round_id=round_id)
+    )
+    if response:
+        return response.get("title")
