+++ openapi/api.yml
@@ -62,7 +62,7 @@ paths:
           required: false
           explode: false
         - in: query
-          name: include_forms
+          name: forms
           style: form
           schema:
             type: boolean
