+++ openapi/components.yml
@@ -22,6 +22,11 @@ components:
           type: string
         last_edited:
           type: string
+        round_name:
+          type: string
+        forms:
+          type: object
+          additionalProperties: true
         date_submitted:
           type: string
     PostApplication:
