+++ pytest.ini
@@ -2,5 +2,5 @@
 env =
     FLASK_ENV=unit_test
     FLASK_DEBUG=1
-    D:DATABASE_URL=postgresql://postgres:postgres@127.0.0.1:5432/fsd_app_store_test
+    D:DATABASE_URL=postgresql://postgres:postgres@127.0.0.1:5433/postgres
 mocked-sessions=db.db.session
