+++ requirements-dev.in
@@ -9,4 +9,3 @@ pytest-flask
 pip-tools
 asserts
 pytest-flask-sqlalchemy
-psycopg2
