+++ requirements.in
@@ -1,6 +1,6 @@
 connexion
 flask_restx
-funding-service-design-utils==1.0.10,<1.1.0
+funding-service-design-utils
 openapi-spec-validator
 prance
 requests
@@ -14,7 +14,7 @@ colored
 #   Database
 #-----------------------------------
 SQLAlchemy
-Flask-SQLAlchemy
+Flask-SQLAlchemy==2.5.1
 Flask-Migrate
 sqlalchemy-utils
 sqlalchemy_json
