+++ requirements.txt
@@ -2,19 +2,21 @@
 # This file is autogenerated by pip-compile with python 3.10
 # To update, run:
 #
-#    pip-compile requirements.in
+#    pip-compile
 #
-alembic==1.8.0
+alembic==1.8.1
     # via flask-migrate
 aniso8601==9.0.1
     # via flask-restx
-attrs==21.4.0
-    # via jsonschema
-babel==2.10.3
+attrs==22.1.0
+    # via
+    #   jsonschema
+    #   openapi-schema-validator
+babel==2.11.0
     # via flask-babel
 blinker==1.5
     # via sentry-sdk
-certifi==2021.10.8
+certifi==2022.9.24
     # via
     #   requests
     #   sentry-sdk
@@ -22,23 +24,23 @@ cffi==1.15.1
     # via cryptography
 chardet==4.0.0
     # via prance
-charset-normalizer==2.0.12
+charset-normalizer==2.1.1
     # via requests
-click==8.0.3
+click==8.1.3
     # via
     #   clickclick
     #   flask
 clickclick==20.10.2
     # via connexion
-colored==1.4.3
+colored==1.4.4
     # via -r requirements.in
 commonmark==0.9.1
     # via rich
-connexion==2.14.0
+connexion==2.14.1
     # via -r requirements.in
-cryptography==37.0.4
+cryptography==38.0.3
     # via pyjwt
-flask==2.1.2
+flask==2.2.2
     # via
     #   connexion
     #   flask-babel
@@ -49,9 +51,9 @@ flask==2.1.2
     #   sentry-sdk
 flask-babel==2.0.0
     # via funding-service-design-utils
-flask-migrate==3.1.0
+flask-migrate==4.0.0
     # via -r requirements.in
-flask-restx==0.5.1
+flask-restx==1.0.3
     # via -r requirements.in
 flask-sqlalchemy==2.5.1
     # via
@@ -59,64 +61,76 @@ flask-sqlalchemy==2.5.1
     #   flask-migrate
 funding-service-design-utils==1.0.10
     # via -r requirements.in
-greenlet==2.0.0
+greenlet==2.0.1
     # via sqlalchemy
 gunicorn==20.1.0
     # via funding-service-design-utils
-idna==3.3
+idna==3.4
     # via requests
+importlib-resources==5.10.0
+    # via openapi-spec-validator
 inflection==0.5.1
     # via connexion
-invoke==1.7.1
+invoke==1.7.3
     # via -r requirements.in
-itsdangerous==2.1.0
+itsdangerous==2.1.2
     # via
     #   connexion
     #   flask
-jinja2==3.0.3
+jinja2==3.1.2
     # via
     #   flask
     #   flask-babel
     #   swagger-ui-bundle
-jsonschema==4.4.0
+jsonschema==4.17.1
     # via
     #   connexion
     #   flask-restx
+    #   jsonschema-spec
     #   openapi-schema-validator
     #   openapi-spec-validator
-mako==1.2.1
+jsonschema-spec==0.1.2
+    # via openapi-spec-validator
+lazy-object-proxy==1.8.0
+    # via openapi-spec-validator
+mako==1.2.4
     # via alembic
-markupsafe==2.1.0
+markupsafe==2.1.1
     # via
     #   jinja2
     #   mako
-openapi-schema-validator==0.2.3
+    #   werkzeug
+openapi-schema-validator==0.3.4
     # via openapi-spec-validator
-openapi-spec-validator==0.4.0
+openapi-spec-validator==0.5.1
     # via -r requirements.in
 packaging==21.3
-    # via connexion
-prance==0.21.8.0
+    # via
+    #   connexion
+    #   prance
+pathable==0.4.3
+    # via jsonschema-spec
+prance==0.22.11.4.0
     # via -r requirements.in
-psycopg2-binary==2.9.3
+psycopg2-binary==2.9.5
     # via -r requirements.in
 pycparser==2.21
     # via cffi
-pygments==2.12.0
+pygments==2.13.0
     # via rich
-pyjwt[crypto]==2.4.0
+pyjwt[crypto]==2.6.0
     # via funding-service-design-utils
 pyparsing==3.0.9
     # via packaging
-pyrsistent==0.18.1
+pyrsistent==0.19.2
     # via jsonschema
 python-dateutil==2.8.2
     # via -r requirements.in
 python-dotenv==0.20.0
     # via funding-service-design-utils
-python-json-logger==2.0.2
+python-json-logger==2.0.4
     # via funding-service-design-utils
-pytz==2022.1
+pytz==2022.6
     # via
     #   babel
     #   flask-babel
@@ -127,31 +141,31 @@ pyyaml==6.0
     #   clickclick
     #   connexion
     #   funding-service-design-utils
+    #   jsonschema-spec
     #   openapi-spec-validator
-requests==2.27.1
+requests==2.28.1
     # via
     #   -r requirements.in
     #   connexion
     #   prance
-rich==12.4.4
+rich==12.6.0
     # via funding-service-design-utils
 ruamel-yaml==0.17.21
     # via prance
-ruamel-yaml-clib==0.2.6
+ruamel-yaml-clib==0.2.7
     # via ruamel-yaml
 semver==2.13.0
     # via prance
-sentry-sdk[flask]==1.9.9
+sentry-sdk[flask]==1.11.1
     # via funding-service-design-utils
 six==1.16.0
     # via
-    #   flask-restx
     #   prance
     #   python-dateutil
     #   sqlalchemy-json
 slugify==0.0.1
     # via -r requirements.in
-sqlalchemy==1.4.39
+sqlalchemy==1.4.44
     # via
     #   -r requirements.in
     #   alembic
@@ -164,11 +178,13 @@ sqlalchemy-utils==0.38.3
     # via -r requirements.in
 swagger-ui-bundle==0.0.9
     # via -r requirements.in
+typing-extensions==4.4.0
+    # via jsonschema-spec
 urllib3==1.26.12
     # via
     #   requests
     #   sentry-sdk
-werkzeug==2.0.3
+werkzeug==2.2.2
     # via
     #   connexion
     #   flask
