+++ tests/conftest.py
@@ -36,7 +36,7 @@ def mock_get_random_choices(
 def mock_get_data_fix(mocker):
     # mock the function in the file it is invoked (not where it is declared)
     mocker.patch(
-        "api.routes.application.helpers.get_data",
+        "external_services.get_data",
         new=mock_get_data,
     )
 
@@ -51,7 +51,7 @@ def mock_random_choices(mocker):
 def mock_post_data_fix(mocker):
     # mock the function in the file it is invoked (not where it is declared)
     mocker.patch(
-        "external_services.http_methods.post_data",
+        "external_services.post_data",
         new=mock_post_data,
     )
