+++ tests/helpers.py
@@ -5,7 +5,33 @@ import urllib
 from typing import List
 
 from config import Config
+from db import db
 from deepdiff import DeepDiff
+from sqlalchemy import func
+from sqlalchemy import inspect
+
+
+def get_random_row(table):
+    """get_random_row Uses a database-side select to get a random row. Does
+    this by using a random offset with range (1, number of rows)
+
+    :param table: Sqlalchemy mapper object
+    :return: A random row from the given mapper.
+    """
+
+    primary_key_name = inspect(table).primary_key[0].name
+    primary_key_column = getattr(table, primary_key_name)
+    return (
+        db.session.query(table)
+        .offset(
+            func.floor(
+                func.random()
+                * db.session.query(func.count(primary_key_column))
+            )
+        )
+        .limit(1)
+        .one()
+    )
 
 
 def local_api_call(endpoint: str, params: dict = None, method: str = "get"):
@@ -197,7 +223,7 @@ application_expected_data = [
     {
         "status": "NOT_STARTED",
         "project_name": "project_name not set",
-        "date_submitted": None,
+        "date_submitted": "null",
         "started_at": "2022-05-20 14:47:12",
         "last_edited": None,
         **application_data,
