+++ tests/test_reports.py
@@ -1,4 +1,5 @@
-from db.models.applications import ApplicationTestMethods
+from db.models import Applications
+from tests.helpers import get_random_row
 from tests.helpers import post_test_applications
 
 
@@ -13,7 +14,7 @@ def test_get_application_statuses(client):
         == b"NOT_STARTED,IN_PROGRESS,SUBMITTED,COMPLETED\r\n3,0,0,0\r\n"
     )
 
-    app = ApplicationTestMethods.get_random_app()
+    app = get_random_row(Applications)
     app.status = "IN_PROGRESS"
 
     response = client.get(
