+++ tests/test_unit_db.py
@@ -1,4 +1,3 @@
-
 def test_joined_app_with_forms():
 
     pass
